+++ SharedFileList.h
@@ -85,7 +85,7 @@ public:
 	CString GetDirNameByPseudo(const CString& strPseudoName) const;
 
 	uint64	GetDatasize(uint64 &pbytesLargest) const;
-	int		GetCount()	{return m_Files_map.GetCount(); }
+	int		GetCount()	{return static_cast<int>(m_Files_map.GetCount()); }
 	int		GetHashingCount()													{ return waitingforhash_list.GetCount()+currentlyhashing_list.GetCount(); }
 	bool	ProbablyHaveSingleSharedFiles() const								{ return bHaveSingleSharedFiles && !m_liSingleSharedFiles.IsEmpty(); } // might not be always up-to-date, could give false "true"s, not a problem currently
