+++ id3lib/src/tag_parse_v1.cpp
@@ -127,7 +127,7 @@ bool id3::v1::parse(ID3_TagImpl& tag, ID3_Reader& reader)
   else
   {
     // trackno[0] != '\0'
-    const int paddingsize = (ID3_V1_LEN_COMMENT-2) - comment.size();
+    const size_t paddingsize = (ID3_V1_LEN_COMMENT-2) - comment.size();
     const char * padding = "                            "; //28 spaces
 
     if (trackno[1] == '\0' || trackno[1] == 0x20 && trackno[0] != 0x20)
